requirements-dev.txt
autopep8
attrs >=17.4
black
bump2version >=1.0.1
check-manifest >=0.36
doc8
flake8
graphviz
invoke >=0.14
ipykernel
ipython >=5.8
isort
jinja2 >= 3.0
m2r2
nbsphinx
numpydoc
pydocstyle
pytest <7.1
pyflakes >=1.4.0
sphinx_book_theme
sphinx_compas_theme >=0.15.18
sphinx ==4.5
sphinx-design
sphinx-inline-tabs
sphinx-togglebutton
sphinx-remove-toctrees
sphinx-copybutton
twine
wheel
-e .
